% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_C.R
\name{calc_C}
\alias{calc_C}
\title{under WAIT discipline calculating system capacity}
\usage{
calc_C(N, P_1, mu_1, mu_2)
}
\arguments{
\item{N}{:= the number of primary servers. No more than N requests for secondary service can exist simultaneously}

\item{P_1}{:= the probability a call requires  secondary service}

\item{mu_1}{:= the primary service rate}

\item{mu_2}{:= the secondary service rate}
}
\value{
C := system capacity (customers/time)
}
\description{
under WAIT discipline calculating system capacity
}
\examples{
N <- 5
P_1 <- .3
mu_1 <- 2
mu_2 <- 4
C <- calc_C(N,P_1, mu_1,mu_2)
}
