% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_E.R
\name{calc_E}
\alias{calc_E}
\title{Under the WAIT discipline, calculating the average total service time}
\usage{
calc_E(mu_p, P_1, W_q)
}
\arguments{
\item{mu_p}{:= the service time of the primary queue}

\item{P_1}{:= the probability the customer needs secondary service}

\item{W_q}{:= the expected value of a secondary customer's service wait time from requesting secondary service until completion of service}
}
\value{
E := the average total service time
}
\description{
Under the WAIT discipline, calculating the average total service time
}
\examples{
mu_p <-5
P_1 <- .3
W_q <- 10
E <- calc_E(mu_p,P_1,W_q)
}
