% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_L_p.R
\name{calc_L_p}
\alias{calc_L_p}
\title{Under the WAIT discipline, calculate the average number of customers}
\usage{
calc_L_p(rho_p, lambda, sigma_sqr)
}
\arguments{
\item{rho_p}{:= the utilization of the primary server}

\item{lambda}{:= the arrival rate of customers (number of customers per unit time)}

\item{sigma_sqr}{:= the service time variance for a all customers}
}
\value{
L_p := the average number of customers in the system
}
\description{
Under the WAIT discipline, calculate the average number of customers
}
\examples{
rho_p <- .6
lambda <- 3
sigma_sqr <- .4
}
