% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_L_Q.R
\name{calc_L_Q}
\alias{calc_L_Q}
\title{under the WAIT discipline, calculating he average number of secondary customers in the system}
\usage{
calc_L_Q(N, rho_Q)
}
\arguments{
\item{N}{:= the number of primary servers}

\item{rho_Q}{:= the utilization of secondary servers}
}
\value{
$$L_Q$$ := the average number of secondary customers in the system
}
\description{
under the WAIT discipline, calculating he average number of secondary customers in the system
}
\examples{
N <- 5
P_1 <- .2
lambda <- 3
mu_2 <- 4
rho_Q <- calc_rho_Q(N, P_1, lambda, mu_2)
L_Q <- calc_L_Q(N, rho_Q)
}
