% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rho_p.R
\name{calc_rho_p}
\alias{calc_rho_p}
\title{Under the WAIT discipline, calculate the utilization of the primary server}
\usage{
calc_rho_p(lambda, E)
}
\arguments{
\item{lambda}{:= the arrival rate of customers (number of customers per unit time)}

\item{E}{:= the average total service time}
}
\value{
rho_p := the utilization of the primary server
}
\description{
Under the WAIT discipline, calculate the utilization of the primary server
}
\examples{
lambda <- 2
E <- .2
rho_p <- calc_rho_p (lambda,E)
}
