% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rho_Q.R
\name{calc_rho_Q}
\alias{calc_rho_Q}
\title{under the WAIT discipline, calculating the secondary server utilization}
\usage{
calc_rho_Q(N, P_1, lambda, mu_2)
}
\arguments{
\item{N}{:=the number of primary servers. No more than N requests for secondary service can exist simultaneously}

\item{P_1}{:= the probability a call requires  secondary service}

\item{lambda}{:= the customer arrival rate}

\item{mu_2}{:= the secondary service rate}
}
\value{
$$rho_Q$$ := utilization of a secondary server. Note: this is rho_s in the original text
}
\description{
under the WAIT discipline, calculating the secondary server utilization
}
\examples{
N <- 5
P_1 <- .2
lambda <- 3
mu_2 <- 4
rho_Q <- calc_rho_Q(N, P_1, lambda, mu_2)
}
