% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sigma_sqr.R
\name{calc_sigma_sqr}
\alias{calc_sigma_sqr}
\title{Under the WAIT discipline, calculating  the variance of the total queue}
\usage{
calc_sigma_sqr(W_q, P_1, sigma_Q_q_sqr, sigma_p_sqr)
}
\arguments{
\item{W_q}{:= the expected value of a secondary customer's service wait time from requesting secondary service until completion of service. Note this is W_s in the original paper.}

\item{P_1}{:= the probability the customer needs secondary service}

\item{sigma_Q_q_sqr}{:= The conditional variance of the wait for secondary service. Note: this is (sigma_s)^2 in the original paper}

\item{sigma_p_sqr}{:= The variance of the primary service time. Note: this is (sigma_1)^2 in the original paper}
}
\value{
(sigma)^2 := the service time variance
}
\description{
Under the WAIT discipline, calculating  the variance of the total queue
}
\examples{
N <- 5
P_1 <- .2
lambda <- 3
mu_1 <- 6
mu_2 <- 4
mu_q <- mu_2
sigma_2_sqr  <- .04
rho_Q <- calc_rho_Q(N, P_1, lambda, mu_2)
L_Q <- calc_L_Q(N, rho_Q)
theta <- calc_theta(rho_Q,N)
W_q <- calc_W_Q(L_Q, N, P_1, lambda)
sigma_Q_q_sqr <- calc_sigma_Q_q_sqr(N,sigma_2_sqr, mu_q,rho_Q, theta)
sigma_p_sqr <- 1/(mu_1^2)
}
