% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sigma_z_sqr.R
\name{calc_sigma_z_sqr}
\alias{calc_sigma_z_sqr}
\title{Under the WAIT discipline, intermediate function for calculating \eqn{\sigma_{Qq}^2}}
\usage{
calc_sigma_z_sqr(N, z, rho_Q, theta)
}
\arguments{
\item{N}{:= the number of primary servers. No more than N requests for secondary service can exist simultaneously}

\item{z}{:= a second intermediate value in calculating \eqn{\sigma_q^2}. Note this is q in the original paper.}

\item{rho_Q}{:= secondary server utilization}

\item{theta}{:= the singular positive root of the polynomial \eqn{\sum_{n=0}^{n=N} \frac{(1-\rho_q)*\theta^n}{n!} = 1}}
}
\value{
\eqn{\sigma_z^2} := the variance of the secondary queue's wait time. Note this is \eqn{\sigma_q^2} in the original paper.
}
\description{
Under the WAIT discipline, intermediate function for calculating \eqn{\sigma_{Qq}^2}
}
\examples{
N <- 5
P_1 <- .2
lambda <- 3
mu_2 <- 4
rho_Q <- calc_rho_Q(N, P_1, lambda, mu_2)
theta <- calc_theta(rho_Q,N)
z <- calc_z(N,rho_Q, theta)
sigma_z_sqr  <- calc_sigma_z_sqr(N,z, rho_Q, theta)
}
