% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_theta.R
\name{calc_theta}
\alias{calc_theta}
\title{Under the WAIT discipline, calculate \eqn{\theta}}
\usage{
calc_theta(rho_Q, N)
}
\arguments{
\item{rho_Q}{:=  utilization of the secondary server. Note: this is rho_s in the original paper.}

\item{N}{:= the number of primary servers. No more than N requests for secondary service can exist simultaneously}
}
\value{
\eqn{\theta} :=the singular positive root of the polynomial \eqn{\sum_{n=0}^{n=N} \frac{(1-\rho_q)*\theta^n}{n!} = 1}
}
\description{
Under the WAIT discipline, calculate \eqn{\theta}
}
\examples{
N <- 5
P_1 <- .2
lambda <- 3
mu_2 <- 4
rho_Q <- calc_rho_Q(N, P_1, lambda, mu_2)
theta <- calc_theta(rho_Q,N)
}
