% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_W_p.R
\name{calc_W_p}
\alias{calc_W_p}
\title{Under the WAIT discipline, calculate the average primary service wait time}
\usage{
calc_W_p(L_p, lambda)
}
\arguments{
\item{L_p}{:=  the average number of customers in the system}

\item{lambda}{:= the arrival rate of customers (number of customers per unit time)}
}
\value{
W_p := the average primary service wait time (in unit time)
}
\description{
Under the WAIT discipline, calculate the average primary service wait time
}
\examples{
lambda <- 2
L_p <- 4
W_p <- calc_W_p(L_p, lambda)
}
