% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_W_P_q.R
\name{calc_W_p_q}
\alias{calc_W_p_q}
\title{Under the WAIT discipline, Calculate the expected delay for the primary service}
\usage{
calc_W_p_q(W_p, E)
}
\arguments{
\item{W_p}{:= the average primary service wait time}

\item{E}{:= the average total service time}
}
\value{
W_p_q := the expected delay for the primary service. Note: this is D_p in the original paper
}
\description{
Under the WAIT discipline, Calculate the expected delay for the primary service
}
\examples{
W_p <- 12
E <- 50
W_p_q <- calc_W_p_q(W_p,E)
}
