% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_system.R
\name{check_system}
\alias{check_system}
\title{Check to make sure the inputs are valid}
\usage{
check_system(lambda, P_1, mu_2, N)
}
\arguments{
\item{lambda}{:= customer arrival rate.}

\item{P_1}{:= probability a customer requires secondary service.}

\item{mu_2}{:= secondary service rate.}

\item{N}{:= the number of primary servers.}
}
\value{
text reporting success of test
}
\description{
Check to make sure the inputs are valid
}
\examples{
lambda <- 2
P_1 <- .2
mu_2 <- 4
N <- 5
check_system(lambda, P_1,mu_2,N)
}
