% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facto.R
\name{facto}
\alias{facto}
\title{compute the factorial}
\usage{
facto(no)
}
\arguments{
\item{no}{:=  the number to take the factorial of}
}
\value{
the factorial of no, i.e. no!
}
\description{
compute the factorial
}
\examples{
val <- facto(3) #val is equal to 3!
}
