% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxn.R
\name{fxn}
\alias{fxn}
\title{Under the WAIT discipline, intermediate function for calculating \eqn{\sigma_{Q_q}^2}}
\usage{
fxn(n, N, rho_Q, theta)
}
\arguments{
\item{n}{:= the number to evaluate the probability distribution at. 0<= n<= N, where no more than N requests can exist simultaneously}

\item{N}{:= the number of primary servers. No more than N requests for secondary service can exist simultaneously}

\item{rho_Q}{:= secondary server utilization}

\item{theta}{:= the singular positive root of the polynomial \eqn{\sum_{n=0}^{n=N} \frac{(1-\rho_q)*\theta^n}{n!} = 1}}
}
\value{
intermediate value in calculating \eqn{\sigma_{Q_q}^2}
}
\description{
Under the WAIT discipline, intermediate function for calculating \eqn{\sigma_{Q_q}^2}
}
\examples{
i <- 1
N <- 5
mu_q  <- 3
P_1 <- .2
lambda <- 1
mu_2 <- mu_q
rho_Q <- calc_rho_Q(N, P_1, lambda, mu_2)
theta <- calc_theta(rho_Q,N)
fxn(i,N, rho_Q, theta)
}
