% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hqueue.R
\name{h_queue}
\alias{h_queue}
\title{Calculation of Hierarchical Queue outputs from Willemain (1974).}
\usage{
h_queue(
  N,
  lambda,
  P_1,
  mu_1,
  mu_2,
  sigma_1 = sqrt(1/(mu_1^2)),
  sigma_2 = sqrt(1/(mu_2^2))
)
}
\arguments{
\item{N}{:= number of primary servers}

\item{lambda}{:customer arrival rate}

\item{P_1}{:= percentage of customers that rwuqire secondary service}

\item{mu_1}{:= primary service rate}

\item{mu_2}{:= secondary service rate}

\item{sigma_1}{:= standard deviation of primary service rate, default is for exponential distribution}

\item{sigma_2}{:= standard deviation of secondary service rate, default is for exponential distribution}
}
\value{
A table of the queue outputs in a dataframe
}
\description{
Calculation of Hierarchical Queue outputs from Willemain (1974).
}
\examples{
lambda <- 2
P_1 <- .2
mu_1 <- 4
sigma_1 <- 1
mu_2 <- 4
sigma_2 <- 1
N <- 5
results <- h_queue(N,lambda, P_1,mu_1, mu_2, sigma_1, sigma_2)
}
