% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_system.R
\name{check_system}
\alias{check_system}
\title{Check to make sure the inputs are valid}
\usage{
check_system(rho_q, rho_p)
}
\arguments{
\item{rho_q}{utilization of secondary servers}

\item{rho_p}{utilization of primary servers}
}
\value{
text reporting success of test
}
\description{
Check to make sure the inputs are valid
}
\examples{
rho_q <- .4
rho_p <- .6
check_system(rho_q, rho_p)
}
